Fix codecheck warnings
Issue: https://gitee.com/openharmony/arkcompiler_runtime_core/issues/ICJSQ7

+++ static_core/compiler/optimizer/code_generator/fast_divisor.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Huawei Device Co., Ltd.
+ * Copyright (c) 2024-2025 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -67,7 +67,7 @@ public:
             delta = ad - r2;
         } while (q1 < delta || (q1 == delta && r1 == 0));
 
-        magic_ = q2 + 1U;
+        magic_ = static_cast<int64_t>(q2) + 1U;
         if (divisor < 0) {
             magic_ = -magic_;
         }
